Fixes
Implemented some hit feedback. After arrow is released there is a screen shake.
Updated my trajectory line, it is not that visable and the more you hold, the more it becomes transparent.
Fixed some accuracy, froze enemis pozition on Z axis, after they die - disabled theri colliders and set their rigibody to kinematic.
Arrow now has a trail, and if the arrow is released with enough strength - it gets special trail and some lighting at the tip of its head.